requirements.txt
h5py==3.7.0
huggingface-hub==0.24.2
imageio==2.26.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipython==8.8.0
keras==2.12.0
matplotlib==3.6.3
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.14.0
opencv-python==4.7.0.68
pillow==10.3.0
protobuf==3.20.3
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
setuptools==44.0.0
tensorboard==2.12.3
tensorflow==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-gan==2.1.0
tensorflow-hub==0.13.0
tensorflow-io==0.34.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
tensorflow-probability==0.19.0
tf==1.13.2
tf-slim==1.1.0
tf2-py==0.7.6
tf2-ros==0.7.6
tqdm==4.65.0
urllib3==1.26.14
wheel==0.38.4
xacro==1.14.14

